Create separate files for tests
