+++ src/tests_metric.c
@@ -1,0 +1,96 @@
+#include "tests.h"
+
+void test_metric_equal_graphs(int* passed, int* failed) {
+    int size = 15;
+    matrix* g1 = matrix_init(size);
+    graph_generate(g1, 7, 1, 0.8f, 1);
+    matrix* g2 = matrix_clone(g1);
+
+    int d = distance(g1, g2);
+
+    graph_print(g1, "Graph 1");
+    graph_print(g2, "Graph 2");
+    printf("\nDistance: %d", d);
+
+    if (d == 0) {
+        print_test_pass(__func__);
+        (*passed)++;
+    }
+    else {
+        print_test_fail(__func__);
+        (*failed)++;
+    }
+
+    printf("\033[0;37m\n\n");
+
+    matrix_destroy(g1);
+    matrix_destroy(g2);
+}
+
+void test_metric_permuted_graphs(int* passed, int* failed) {
+    int size = 20;
+    matrix* g1 = matrix_init(size);
+    graph_generate(g1, 1, 0, 0.8f, 1);
+    matrix* g2 = matrix_clone(g1);
+    graph_permute(g2);
+
+    graph_print(g1, "Graph 1");
+    graph_print(g2, "Graph 2");
+
+    int d = distance(g1, g2);
+
+    printf("\nDistance: %d", d);
+
+    if (d == 0) {
+        print_test_pass(__func__);
+        (*passed)++;
+    }
+    else {
+        print_test_fail(__func__);
+        (*failed)++;
+    }
+
+    printf("\033[0;37m\n\n");
+
+    matrix_destroy(g1);
+    matrix_destroy(g2);
+}
+
+void test_metric_different_graphs(int* passed, int* failed) {
+    int size = 11;
+    matrix* g1 = matrix_init(size);
+    graph_generate(g1, 7, 1, 0.8f, 1);
+    matrix* g2 = matrix_clone(g1);
+
+    graph_add_noise(g2, 0.5f, 1, 0.0);
+
+    int d = distance(g1, g2);
+
+    graph_print(g1, "Graph 1");
+    graph_print(g2, "Graph 2");
+    printf("\nDistance: %d", d);
+
+    if (d != 0) {
+        print_test_pass(__func__);
+        (*passed)++;
+    }
+    else {
+        print_test_fail(__func__);
+        (*failed)++;
+    }
+
+    printf("\033[0;37m\n\n");
+
+    matrix_destroy(g1);
+    matrix_destroy(g2);
+}
+
+
+void tests_metric(int* passed, int* failed) {
+
+    printf("Running metrics tests...\n");
+
+    test_metric_equal_graphs(passed, failed);
+    test_metric_permuted_graphs(passed, failed);
+    test_metric_different_graphs(passed, failed);
+}
